Restructure frontend code; Remove unused code;
Cleanup

Delete unused script and style files

Refactored frontend code folder structure

Remove unnecessary script; Normalize style and script paths

Cleanup ajaxService; Refactor popupFactory exports and their imports
