+++ forum/qa-include/pages/question-post.php
@@ -19,7 +19,6 @@
 
 	More about this license: http://www.question2answer.org/license.php
 */
-//var_dump('hello???');
 
 	if (!defined('QA_VERSION')) { // don't allow this page to be requested directly from browser
 		header('Location: ../');
