+++ forum/qa-plugin/report-reason/frontend/.prettierrc
