+++ forum/qa-plugin/report-reason/frontend/src/misc.js
