+++ forum/qa-plugin/report-reason/frontend/style.css
