+++ forum/qa-plugin/report-reason/q2apro-flag-reasons-layer.php
@@ -18,89 +18,17 @@ class qa_html_theme_layer extends qa_html_theme_base
 
             $flagReasonsMetadata = json_encode(array_merge($flagReasonNoticeLength, $reportFlagList, $reportPopupLabels));
 
-            $this->output('
+            $this->output(
+            '
+                <link rel="stylesheet" href="' . QA_HTML_THEME_LAYER_URLTOROOT . 'frontend/style.css">
                 <script>
                     const FLAG_REASONS_METADATA = Object.freeze(' . $flagReasonsMetadata . ');
                 </script>
-                <script type="text/javascript" src="' . QA_HTML_THEME_LAYER_URLTOROOT . '/dist/script.js"></script>
-                <link rel="stylesheet" href="' . QA_HTML_THEME_LAYER_URLTOROOT . 'style.css">
+                <script type="text/javascript" src="' . QA_HTML_THEME_LAYER_URLTOROOT . 'frontend/dist/script.js"></script>
             '
             );
         }
     }
-
-//    public function q_view_buttons($q_view)
-//    {
-//        if ($this->checkPostData($q_view)) {
-////            $q_view['form']['buttons']['flag']['tags'] =
-////                'data-postid="' . $q_view['raw']['postid'] . '" data-posttype="q" ';
-//        }
-//        parent::q_view_buttons($q_view);
-//    }
-
-//    public function a_item_buttons($a_item)
-//    {
-//        if ($this->checkPostData($a_item)) {
-////            $a_item['form']['buttons']['flag']['tags'] =
-////                'data-postid="' . $a_item['raw']['postid'] . '" data-posttype="a" ';
-//        }
-//        parent::a_item_buttons($a_item);
-//    }
-
-//    public function c_item_buttons($c_item)
-//    {
-//        if ($this->checkPostData($c_item)) {
-////            $c_item['form']['buttons']['flag']['tags'] = 'data-postid="' . $c_item['raw']['postid'] . '" data-posttype="c" data-parentid="' . $c_item['raw']['parentid'] . '" ';
-//        }
-//        parent::c_item_buttons($c_item);
-//    }
-
-//    public function body_hidden()
-//    {
-//        if ($this->isLogged && 'question' === $this->template) {
-//            $this->output(
-//                '
-//            <div id="flagbox-popup" class="modal-background flagbox-popup" hidden>
-//                <div class="flagbox-center">
-//                    <div class="qa-flag-reasons-wrap">
-//                        <p><b>
-//                          '
-//                          . qa_lang('q2apro_flagreasons_lang/reason')
-//                          . '
-//                        </b></p>
-//                        ');
-//                        for ($i=1; $i<=6; $i++) {
-//                            $this->output('<label for="qa-spam-reason-radio-' . $i . '">
-//                            <input type="radio" class="qa-spam-reason-radio" name="qa-spam-reason-radio" id="qa-spam-reason-radio-' . $i . '" value="' . $i . '">
-//                            '
-//                            . q2apro_flag_reasonid_to_readable($i)
-//                            . '
-//                            </label>');
-//                        }
-//                        $this->output('
-//                        <p><b>
-//                            '
-//                            . qa_lang('q2apro_flagreasons_lang/note')
-//                            . '
-//                        </b></p>
-//                        <div class="qa-spam-reason-text-wrap">
-//                            <input type="text" name="qa-spam-reason-text" class="qa-spam-reason-text" placeholder="'
-//                . qa_lang('q2apro_flagreasons_lang/enter_details')
-//                . '">
-//                        <div id="qa-spam-reason-error" class="qa-error" hidden></div></div>
-//
-//                        <input type="button" class="qa-form-tall-button qa-form-tall-button-ask qa-form-wide-text qa-go-flag-send-button" value="'
-//                . qa_lang('q2apro_flagreasons_lang/send')
-//                . '">
-//
-//                        <button class="close-preview-btn">X</button>
-//                    </div>
-//                </div>
-//            </div>
-//            ');
-//        }
-//        parent::body_hidden();
-//    }
 
     public function post_meta_flags($post, $class)
     {
@@ -115,7 +43,6 @@ class qa_html_theme_layer extends qa_html_theme_base
                 unset($post['flags']);
 
                 $post['flags']['data'] = self::prepareFlagSuffix($flagsCount);
-//                    $post['flags']['suffix'] = ': <br>' . $flagInfo;
                 $post['flags']['suffix'] = $flagInfo;
             }
         }
